Cast H3 sampler state to diffusion precision
Convert packed float32 sampler state to the configured H3 diffusion compute dtype before model evaluation, matching the reference model wrapper while keeping float32 velocity output for the solver.\n\nReviewed: both local and sequence evaluation paths share the cast, and focused runtime coverage checks the model-facing dtype.

Amp-Thread-ID: https://ampcode.com/threads/T-01a0c18f-755f-74ff-87ee-b869c9cbfb31

+++ packages/dinkster-inference-torch/src/dinkster_inference_torch/minimax_h3_runtime.py
@@ -1533,7 +1533,7 @@ class MiniMaxH3DiTRuntime(MultiStreamSamplingRuntime):
                     inner_kernel = None
                     sequence_group_ranks: tuple[int, ...] = ()
                     try:
-                        av = unpack_latent_streams(x, layout)
+                        av = unpack_latent_streams(x.to(dtype=compute_dtype), layout)
                         facts = _packed_sequence_facts(
                             av,
                             context,
@@ -1687,7 +1687,7 @@ class MiniMaxH3DiTRuntime(MultiStreamSamplingRuntime):
                         sequence_sharding=sharding,
                     )
                 else:
-                    av = unpack_latent_streams(x, layout)
+                    av = unpack_latent_streams(x.to(dtype=compute_dtype), layout)
                     if attention_kernel_factory is None:
                         velocity = model(
                             av,
